controller: Prepare structure around CT zone limiting.
In order to be able to store CT limits for specified zone, store the
zone inside separate struct instead of simap. This allows to add
the addition of limit without changing the whole infrastructure again.

This is a preparation step for the CT zone limits.

Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.
    Address comments from Lorenzo.
v3: Rebase on top of latest main.
v2: Fix NULL ptr deref.