northd: Add LR option to commit all traffic.
Add LR option to commit all traffic that is not already committed
by either NAT or LB. This ensures that the traffic is tracked, and
we don't erroneously commit reply traffic, or reply traffic is not
marked as invalid.

To achieve the commit we need to perform lookup on every packet
that goes through LR pipeline whenever there is stateful NAT.

The SNAT lookup requires additional flag as the unSNAT is happening
in ingress pipeline and at that point we need to know if the packet
is reply or not. This is not required for DNAT, because unDNAT stage
happens in egress.

This also helps with HWOL as there shouldn't be any match on ct.new
for established sessions as we will commit everything in addition to
already existing stateful NATs and LBs.

Reported-at: https://issues.redhat.com/browse/FDP-787
Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>
Fixes: ffe267317c25 ("northd: Don't skip the unSNAT stage for traffic towards VIPs.")
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v6: Rebase on top of latest main.
    Address typos in documentation.
    Avoid smap parsing for each NAT and pass it as argument.
    Add missing documentation about HWOL.
    Add missing Fixes tag.
    Add ack from Mark.
    Add ack from Han.

v5: Rebase on top of latest main.
    Address typos in commit message and documentation.
    Keep the flow ct_commit {} intact if all NATs are stateless.
    Add extra optimization flag when the unSNAT reports ct.new skip additional lookup in egress pipeline.

v4: Rebase on top of latest main.
    Adjust the option name.
    Adjust the NEWS entry.
    Update ovn-northd.8.xml.
    Change the option so it affects both zones regardless if there is only stateful SNAT or DNAT.
    Add comment about !ct.rpl optimization.

v3: Rebase on top of latest main.
    Add extra system test that checks specific scenario which was broken without this option.