inc-engine: Adjust the force recompute API.
There are cases when we need to wake the thread immediately after
setting force recompute. This is mainly in error handling that
happens after engine_run() call. In order to achieve that make
the API more ergonomic with parameters to force the wake if needed.

Reported-at: https://issues.redhat.com/browse/FDP-753
Signed-off-by: Ales Musil <amusil@redhat.com>