physical: Allow l3gateway and patch port to be peers.
Allow the l3gateway and patch port to be peers for connected LRPs,
previously it was only allowed to connect two ports of the same
type. This allows the CMS to connect any router with GW router
which is required for the concept of transit routers.

Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.
v3: Rebase on top of latest main.
    Fix the wrong comparison in I-P handler.
    Add ack from Numan.
v3: Rebase on top of latest main.
    Add function to compare the types.