physical: Allow l3gateway and patch port to be peers.
Allow the l3gateway and patch port to be peers for connected LRPs,
previously it was only allowed to connect two ports of the same
type. This allows the CMS to connect any router with GW router
which is required for the concept of transit routers.

Signed-off-by: Ales Musil <amusil@redhat.com>
---
v2: Rebase on top of current main.
    Add function to compare the types.