northd: Introduce the concept of transit routers.
In order to allow the concept of transit router set the port binding
as "remote" and assign the chassis whenever the requested-chassis
for LRP is remote chassis. The only supported requested-chassis for
now is remote, it might be extended in the future if needed.

The transit router has very similar behavior to distributed router
that it will lose the CT state between chassis, in this case between
AZs.

Reported-at: https://issues.redhat.com/browse/FDP-872
Acked-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.
v3: Rebase on top of latest main.
v2: Rebase on top of latest main.
    Add ack from Lorenzo.
    Adjust the condition to set the type and chassis for PB.