actions, physical: Make the MC split action generic.
The action used for multiple split buffer scenario, when the
action wouldn't fit into the netlink buffer, can be used in
different context than just the multicast groups. Change the
naming and add generic helper so it is clearer that is indeed
the case.

Acked-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.
v3: Rebase on top of latest main.
v2: Rebase on top of latest main.
    Add ack from Lorenzo.