vec: Add new data type called vector.
The aim of the structure is to reduce certain type of allocations
and help with the cache consistency. There are places in the codebase
that essentially do vectoring just raw with x2nrealloc() calls. The
second category are computations that use struct ovs_list to achieve
array like iterations.

This is preparation with the structure testing, the real replacement
is done later on.

Signed-off-by: Ales Musil <amusil@redhat.com>