ci: Add extre distcheck job.
The log directory is removed with each distcheck run. This would
remove logs from failed jobs in the "stable" part of the run. In
order to prevent that run all test just as "check" and add extra
line for the "distcheck" that will not run any test. This also
saves some run time as the distcheck was doing the compilation
again which doesn't have to be done twice.

Signed-off-by: Ales Musil <amusil@redhat.com>