controller: Remove only commited virtual port binding requests.
The virtual port binding requests were commited twice in case there
was a failure in between them. That unfortunately doesn't solve the
issue for very rare case when the commit fails multiple times in
a row. To make sure we remove only really commited requests use
the next_cfg and cur_cfg mechanism of idl_loop. We can store
the cur_cfg and check against the next loop cur_cfg to see
if the commit was successful. This mechanism is exposed to
the whole pinctrl so it can be used by any other SB transaction
if needed.

Signed-off-by: Ales Musil <amusil@redhat.com>