controller: Create physical flows based on EVPN structures.
Add physical flows based on the EVPN bindings and multicast groups
that ensure the proper flow of traffic that is going through the
EVPN. In order for this to wrk the CMS needs to ensure that there
is proper VXLAN tunnel port in br-int and also the
"external_ids:ovn-evpn-local-ip" is set properly in the local OvS
database.

All traffic will be flooded through the UNKNOWN group if there is
unknown port at this stage. That's due to missing support for FDB
learning.

Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v4: Rebase on top of latest main.
    Add Xavier's ack.
    Address review comments:
    - Fix default value for local IP.
    - Fix indentation and spelling.

v3: Rebase on top of latest main.
    Add an assert that mc groups have some bindings.

v2: Rebase on top of latest main.
    Add WARN about mossing local_ip.