northd: Add an option to specify EVPN vni in logical switches.
Propagate the VNI option from LS to the datapath, this is the last
missing piece required to activate all the code around EVPN in
ovn-controller. The LS is considered to be part of EVPN when
the "dynamic-routing-vni" is set to valid integer.

Reported-at: https://issues.redhat.com/browse/FDP-1385
Signed-off-by: Ales Musil <amusil@redhat.com>
---
v2: Rebase on top of latest main.
    Add system test.
    Fix typo in documentation.