controller: Switch idle_timeout to hard_timeout for dynamic EVPN FDBs.
The dynamically learned FBDs were using idle_timeout configured to
the same value as fdb_age_threshold for given LS. This could cause
an issue when the traffic was originated only from OVN, without the
reply the flow would be used, but OVN could be just blackholing
traffic.

Switch to hard_timeout instead, that means the flow will be removed
after the specified timeout regardless of the usage, however
this is still better than the blackhole that isn't recoverable.
Without the FDB we would flood instead, which might happen during the
period when hard_timeout expires before the workload replies back to
populate the FDB again.

Fixes: 53e0d5ac2a74 ("controller: Learn EVPN workload FDB from the ingress traffic.")
Signed-off-by: Ales Musil <amusil@redhat.com>